+++ sdks/python/hatchet_sdk/runnables/workflow.py
@@ -244,6 +244,24 @@ class BaseWorkflow(Generic[TWorkflowInput]):
             ),
         )
 
+    def _combine_additional_metadata(
+        self, additional_metadata_from_trigger: JSONSerializableMapping
+    ) -> JSONSerializableMapping:
+        return {
+            **self.config.default_additional_metadata,
+            **additional_metadata_from_trigger,
+        }
+
+    def _create_options_with_combined_additional_meta(
+        self, options: TriggerWorkflowOptions
+    ) -> TriggerWorkflowOptions:
+        options_copy = options.model_copy()
+        options_copy.additional_metadata = self._combine_additional_metadata(
+            options.additional_metadata
+        )
+
+        return options_copy
+
     @property
     def input_validator(self) -> type[TWorkflowInput]:
         return cast(type[TWorkflowInput], self.config.input_validator)
@@ -285,7 +303,7 @@ class BaseWorkflow(Generic[TWorkflowInput]):
         return WorkflowRunTriggerConfig(
             workflow_name=self.config.name,
             input=self._serialize_input(input),
-            options=options,
+            options=self._create_options_with_combined_additional_meta(options),
             key=key,
         )
 
@@ -620,7 +638,7 @@ class Workflow(BaseWorkflow[TWorkflowInput]):
         return self.client._client.admin.run_workflow(
             workflow_name=self.config.name,
             input=self._serialize_input(input),
-            options=options,
+            options=self._create_options_with_combined_additional_meta(options),
         )
 
     def run(
@@ -642,7 +660,7 @@ class Workflow(BaseWorkflow[TWorkflowInput]):
         ref = self.client._client.admin.run_workflow(
             workflow_name=self.config.name,
             input=self._serialize_input(input),
-            options=options,
+            options=self._create_options_with_combined_additional_meta(options),
         )
 
         return ref.result()
@@ -665,7 +683,7 @@ class Workflow(BaseWorkflow[TWorkflowInput]):
         return await self.client._client.admin.aio_run_workflow(
             workflow_name=self.config.name,
             input=self._serialize_input(input),
-            options=options,
+            options=self._create_options_with_combined_additional_meta(options),
         )
 
     async def aio_run(
@@ -686,7 +704,7 @@ class Workflow(BaseWorkflow[TWorkflowInput]):
         ref = await self.client._client.admin.aio_run_workflow(
             workflow_name=self.config.name,
             input=self._serialize_input(input),
-            options=options,
+            options=self._create_options_with_combined_additional_meta(options),
         )
 
         return await ref.aio_result()